Skip making new vector when mock coll already vec
9edd10d056b5b25140eaf615f6d0351dd73384fc itself had a perf bug since
`vec` isn't fully _performance_ idempotent.  If given a vector, `vec`
will return a new vector, which isn't the behavior we want, we just
want to return the existing vector.  This commit fixes the issue
which actually delivers the constant-time performance on `rpush`.

Actually tested against benchmarks with this literal code, this time,
not code that was similar-but-not-identical, and verified this is it.